menuitems are now added in baristaresource and using events they are consumed by orderservice which gets a copy of menuitems data from barista (cqrs pattern). all events are now either commands or events (interfaces) createorderrequest and response DTO are now used for placing order

+++ orders/src/main/java/com/fisnikz/coffee_express/orders/entity/OrderItem.java
@@ -1,11 +1,11 @@
 package com.fisnikz.coffee_express.orders.entity;
 
-import javax.persistence.Embeddable;
+import javax.persistence.*;
 
 @Embeddable
 public class OrderItem {
 
-    public int menuItemId;
+    public long menuItemId;
     public String menuItemName;
     public double menuItemPrice;
     public short quantity;
@@ -13,7 +13,7 @@ public class OrderItem {
     public OrderItem() {
     }
 
-    public OrderItem(int menuItemId, String menuItemName, double menuItemPrice, short quantity) {
+    public OrderItem(long menuItemId, String menuItemName, double menuItemPrice, short quantity) {
         this.menuItemId = menuItemId;
         this.menuItemName = menuItemName;
         this.menuItemPrice = menuItemPrice;
